pyproject.toml
[tool.poetry]
name = "data-model-template"
version = "0.1.0"
description = ""
authors = ["charles beniac"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
fastapi = "^0.110.2"
sqlalchemy = "^2.0.29"
scikit-learn = "^1.4.2"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
pytest = "^8.1.1"
pydantic = "^2.7.1"
uvicorn = "^0.29.0"
seaborn = "^0.13.2"
streamlit = "^1.33.0"
mlflow = "^2.12.1"
jwt = "^1.3.1"
azureml-mlflow = "^1.55.0"
azureml-core = "^1.55.0.post2"
asyncio = "^3.4.3"
setuptools = "^69.5.1"
watchdog = "^4.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"
autopep8 = "^2.1.0"
django = "^5.0.6"
whitenoise = "^6.6.0"
pytest-django = "^4.8.0"
opentelemetry-instrumentation-django = "^0.46b0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-api = "^1.25.0"
opentelemetry-exporter-otlp = "^1.25.0"
psycopg2-binary = "^2.9.9"
azure-monitor-opentelemetry = "^1.6.0"
azure-monitor-opentelemetry-exporter = "^1.0.0b26"
mkdocs = "^1.6.0"
mkdocstrings = "^0.25.1"
mkdocstrings-python = "^1.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
